Move CLI support modules out of bin
